# staticServer/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sense100 接口文档",
        "title": "sense100",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "740640431@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.5"
    },
    "host": "192.168.1.104:9494",
    "basePath": "/",
    "paths": {
        "/carrier/createOrUpdateCarrier": {
            "post": {
                "description": "创建承运商",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carrier"
                ],
                "summary": "创建承运商",
                "parameters": [
                    {
                        "description": "createOrUpdateCarrierReq",
                        "name": "createOrUpdateCarrierReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/request.CreateOrUpdateCarrierReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/carrier/createOrUpdateSetMeal": {
            "post": {
                "description": "创建和更新套餐",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carrier"
                ],
                "summary": "创建和更新套餐",
                "parameters": [
                    {
                        "description": "ssSetMeal",
                        "name": "ssSetMeal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsSetMeal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/carrier/getSetMealInfo/{carrierId}/{mouldId}": {
            "get": {
                "description": "根据承运商Id和模板Id获取套餐信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carrier"
                ],
                "summary": "根据承运商Id和模板Id获取套餐信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "carrierId",
                        "name": "carrierId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "mouldId",
                        "name": "mouldId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.BaseResponseSwag"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/common/getSubsetRegion/{parentId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "获取子集城市集合",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "parentId",
                        "name": "parentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SsRegion"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/company/createOrUpdateFacade": {
            "post": {
                "description": "添加或更新企业门面信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "添加或更新企业门面信息",
                "parameters": [
                    {
                        "description": "ssCompanyFacadeInfo",
                        "name": "ssCompanyFacadeInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsCompanyFacadeInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/company/createOrUpdateTrade": {
            "post": {
                "description": "创建或更新行业类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "创建或更新行业类型",
                "parameters": [
                    {
                        "description": "trade",
                        "name": "trade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsTrade"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/company/getDynamicListPage": {
            "post": {
                "description": "分页获取动态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "分页获取动态",
                "parameters": [
                    {
                        "description": "getDynamicListPageReq",
                        "name": "getDynamicListPageReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/request.GetDynamicListPageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DynamicResp"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/company/getFacadeOnUserId/{userId}": {
            "get": {
                "description": "根据用户ID获取企业门面信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "根据用户ID获取企业门面信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.CompanyFacadeInfoResp"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/company/getTradesOnParentId/{parentId}": {
            "get": {
                "description": "根据父级Id获取行业类型集合",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "根据父级Id获取行业类型集合",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "parentId",
                        "name": "parentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SsTrade"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/company/publishDynamic": {
            "post": {
                "description": "发布动态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "发布动态",
                "parameters": [
                    {
                        "description": "ssSpDynamic",
                        "name": "ssSpDynamic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.SsSpDynamic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
        "/order/submitOrder": {
            "post": {
                "description": "提交订单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "提交订单",
                "parameters": [
                    {
                        "description": "submitOrderReq",
                        "name": "submitOrderReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/request.SubmitOrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.BaseResponseSwag"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.SsCompanyFacadeInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "企业地址",
                    "type": "string"
                },
                "bannerUrls": {
                    "description": "bannerUrl,逗号分割",
                    "type": "string"
                },
                "businessBeginTime": {
                    "description": "营业开始时间",
                    "type": "string"
                },
                "businessEndTime": {
                    "description": "营业结束时间",
                    "type": "string"
                },
                "companyId": {
                    "description": "关联企业Id",
                    "type": "integer"
                },
                "createdTime": {
                    "type": "string"
                },
                "creatorId": {
                    "type": "integer"
                },
                "enabled": {
                    "type": "integer"
                },
                "expiryTime": {
                    "description": "过期时间",
                    "type": "string"
                },
                "iD": {
                    "type": "integer"
                },
                "latitude": {
                    "description": "纬度",
                    "type": "number"
                },
                "longitude": {
                    "description": "经度",
                    "type": "number"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "pictureUrls": {
                    "description": "图片url",
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "rentDays": {
                    "description": "购买天数",
                    "type": "integer"
                },
                "startUseTime": {
                    "description": "购买，开始使用时间",
                    "type": "string"
                },
                "state": {
                    "description": "状态(-1欠费、0正常)",
                    "type": "integer"
                },
                "telephone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "tradeId": {
                    "description": "关联行业类型Id",
                    "type": "integer"
                },
                "updatedTime": {
                    "type": "string"
                },
                "updatorId": {
                    "type": "integer"
                },
                "userId": {
                    "description": "关联用户Id",
                    "type": "integer"
                },
                "videoUrl": {
                    "description": "宣传视频url",
                    "type": "string"
                },
                "visitorNumber": {
                    "description": "访问量",
                    "type": "integer"
                },
                "wxAccount": {
                    "description": "微信账号",
                    "type": "string"
                }
            }
        },
...
        "model.SsRegion": {
            "type": "object",
            "properties": {
                "cityCode": {
                    "type": "string"
                },
                "iD": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "levelType": {
                    "type": "integer"
                },
                "lng": {
                    "type": "number"
                },
                "mergerName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "integer"
                },
                "pinyin": {
                    "type": "string"
                },
                "postCode": {
                    "type": "string"
                },
                "shortName": {
                    "type": "string"
                }
            }
        },
        "model.SsSetMeal": {
            "type": "object",
            "required": [
                "carrierId",
                "mouldId",
                "name",
                "coverUrl",
                "creatorId"
            ],
            "properties": {
                "carrierId": {
                    "description": "承运商Id",
                    "type": "integer"
                },
                "coverUrl": {
                    "description": "封面url",
                    "type": "string"
                },
                "createdTime": {
                    "type": "string"
                },
                "creatorId": {
                    "type": "integer"
                },
                "description": {
                    "description": "套餐描述",
                    "type": "string"
                },
                "enabled": {
                    "type": "integer"
                },
                "iD": {
                    "type": "integer"
                },
                "mealRoles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SsSetMealRole"
                    }
                },
                "mouldId": {
                    "description": "模版Id",
                    "type": "integer"
                },
                "name": {
                    "description": "套餐名称",
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "updatedTime": {
                    "type": "string"
                },
                "updatorId": {
                    "type": "integer"
                }
            }
        },
...
        "model.SsTrade": {
            "type": "object",
            "required": [
                "parentId",
                "name",
                "creatorId"
            ],
            "properties": {
                "createdTime": {
                    "type": "string"
                },
                "creatorId": {
                    "description": "管理员Id",
                    "type": "integer"
                },
                "enabled": {
                    "type": "integer"
                },
                "iD": {
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "关联父级id",
                    "type": "integer"
                },
                "remarks": {
                    "type": "string"
                },
                "updatedTime": {
                    "type": "string"
                },
                "updatorId": {
                    "description": "管理员Id",
                    "type": "integer"
                }
            }
        },
        "request.CreateOrUpdateCarrierReq": {
            "type": "object",
            "required": [
                "adminId",
                "name",
                "companyName",
                "contactPerson",
                "contactPhone",
                "legalPersonName",
                "legalPersonIdCard",
                "legalPersonPhone",
                "legalPersonEmail",
                "idCardPositivePicUrl",
                "idCardInsteadPicUrl",
                "idCardHeadPositivePicUrl",
                "businessLicensePicUrl",
                "organizationPicUrl"
            ],
            "properties": {
                "adminId": {
                    "description": "管理员Id",
                    "type": "integer"
                },
                "businessLicensePicUrl": {
                    "type": "string"
                },
                "companyName": {
                    "description": "企业名称",
                    "type": "string"
                },
                "contactEmail": {
                    "description": "联系邮箱",
                    "type": "string"
                },
                "contactPerson": {
                    "description": "联系人姓名",
                    "type": "string"
                },
                "contactPhone": {
                    "description": "联系电话",
                    "type": "string"
                },
                "iD": {
                    "type": "integer"
                },
                "idCardHeadPositivePicUrl": {
                    "type": "string"
                },
                "idCardInsteadPicUrl": {
                    "description": "身份证",
                    "type": "string"
                },
                "idCardPositivePicUrl": {
                    "description": "身份证正面照",
                    "type": "string"
                },
                "legalPersonEmail": {
                    "description": "法人联系邮箱",
                    "type": "string"
                },
                "legalPersonIdCard": {
                    "description": "法人身份证",
                    "type": "string"
                },
                "legalPersonName": {
                    "description": "法人姓名",
                    "type": "string"
                },
                "legalPersonPhone": {
                    "description": "法人联系电话",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "organizationPicUrl": {
                    "type": "string"
                }
            }
        },
        "request.GetDynamicListPageReq": {
            "type": "object",
            "properties": {
                "companyId": {
                    "description": "企业Id",
                    "type": "integer"
                },
                "facadeId": {
                    "description": "门面id",
                    "type": "integer"
                },
                "isPage": {
                    "description": "是否分页",
                    "type": "boolean"
                },
                "pageIndex": {
                    "description": "页面索引",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "页面数据个数",
                    "type": "integer"
                },
                "programId": {
                    "description": "小程序Id",
                    "type": "integer"
                },
                "userId": {
                    "description": "用户Id，没有可以不传",
                    "type": "integer"
                }
            }
        },
...
        "response.BaseResponseSwag": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据jsonObject",
                    "type": "string"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                }
            }
        },
...
    }
}